LPS-57375 Move test classes
